專案優化進度
新增
  -landmarks資料表儲存快速定位地標
  -地標載入工具(landmarks.py)取得台北捷運站座標
  -地圖選定快速定位功能
  -「餐廳名冊」頁籤:顯示環保和客家餐廳名冊
  -在footer顯示開放資料集來源(餐廳/快速定位)
  -所有js和css資源檔加入版本參數,避免瀏覽器快取
  -評論標記篩選功能
優化
  -統一元素互動視覺效果
  -餐廳詳細資訊頁面捲動後會顯示餐廳名稱頁首
  -推薦餐廳左右捲動箭頭僅在可用時顯示
  -動態載入評論標記,過程顯示載入中動畫
修復
  -多個不同頁面收藏餐廳不同步的問題(以按鈕當前狀態為主)
  -特定拖曳區域中可以拖曳html元素的問題

+++ 專案檔案/地圖結果串接/web/handler/toggle_favorite.php
@@ -12,6 +12,7 @@
       exit();
     }
     $storeId = $_POST['storeId'];
+    $currentState = $_POST['currentState'];
 
     // 檢查 storeId 是否存在
     $stmt = bindPrepare($conn, "
@@ -24,7 +25,7 @@
     $stmt->close();
 
     if ($storeCount == 0) {
-      echo json_encode(['success' => false, 'message' => '商家不存在']);
+      echo json_encode(['success' => false, 'message' => '餐廳不存在']);
       exit();
     }
 
@@ -39,6 +40,10 @@
     $stmt->close();
 
     if ($count>0) {
+      if ($currentState == '0') {
+        echo json_encode(['success' => true, 'isFavorite' => true, 'refresh' => true, 'message' => '該餐廳已加入至收藏']);
+        exit();
+      }
       $stmt = bindPrepare($conn, "
         DELETE FROM favorites
         WHERE member_id = ? AND store_id = ?
@@ -46,19 +51,23 @@
       if ($stmt->execute()) {
         echo json_encode(['success' => true, 'isFavorite' => false, 'message' => '已將該餐廳從收藏中移除']);
       } else {
-        echo json_encode(['success' => false, 'message' => $conn->error]);
+        echo json_encode(['success' => false, 'refresh' => false, 'message' => $conn->error]);
       }
       $stmt->close();
     } else {
-        $stmt = bindPrepare($conn, "
-          INSERT INTO favorites (member_id, store_id)
-          VALUES (?, ?)
-        ", "ii", $MEMBER_ID, $storeId);
-        if ($stmt->execute()) {
-          echo json_encode(['success' => true, 'isFavorite' => true, 'message' => '已將該餐廳加入收藏']);
-        } else {
-          echo json_encode(['success' => false, 'message' => $conn->error]);
-        }
-        $stmt->close();
+      if ($currentState == '1') {
+        echo json_encode(['success' => true, 'isFavorite' => false, 'refresh' => true, 'message' => '該餐廳已從收藏中移除']);
+        exit();
+      }
+      $stmt = bindPrepare($conn, "
+        INSERT INTO favorites (member_id, store_id)
+        VALUES (?, ?)
+      ", "ii", $MEMBER_ID, $storeId);
+      if ($stmt->execute()) {
+        echo json_encode(['success' => true, 'isFavorite' => true, 'refresh' => false, 'message' => '已將該餐廳加入收藏']);
+      } else {
+        echo json_encode(['success' => false, 'refresh' => false, 'message' => $conn->error]);
+      }
+      $stmt->close();
     }
   }
